🚚 Rename function to lowercase
