+++ functions/index.js
@@ -4,6 +4,6 @@ if (!process.env.K_SERVICE || process.env.K_SERVICE === 'ssrapp') {
   exports.ssrapp = require('./ssrapp');
 }
 
-if (!process.env.K_SERVICE || process.env.K_SERVICE === 'superLikeRedirect') {
-  exports.superLikeRedirect = require('./superLikeRedirect');
+if (!process.env.K_SERVICE || process.env.K_SERVICE === 'superlike') {
+  exports.superlike = require('./superlike');
 }
